[backend/frontend] Refacto of forms domain and bug fix in front (#15147)
